pyproject.toml
[project]
name = "narramind"
version = "0.1.0"
description = "NarraMind：智能角色扮演与记忆管理系统"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.1.0",
    "mysql-connector-python>=8.0.0",
    "python-dotenv>=0.19.0",
    "uuid>=1.30",
]
